Revised UserUI for web-based implementation
UserUI has been fixed to no longer use any terminal-based interaction!
This involved fixing some things in Person and PersonHome, for the sake
of good abstraction. Driver.java stopped comiling because of removed
methods, so I just took out the method calls. So it compiles, but it's
broken. So don't use it :P